requirements.txt
django==2.0.3
django-webpack-loader==0.6.0
channels_redis==2.1.1
channels==2.0.2
djangorestframework==3.7.7
pycrypto==2.6.1